requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
dj-stripe==2.8.3
Django==4.2.6
django-allauth==0.58.1
django-bitfield==2.2.0
django-countries==7.5.1
django-filter==23.5
django-geo==0.8
django-js-asset==2.2.0
django-leaflet==0.29.0
django-mptt==0.15.0
django-rest-framework==0.1.0
django-separatedvaluesfield==0.6.0
django-timezone-field==6.1.0
django-uuidfield==0.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docopt==0.6.2
GDAL @ file:///C:/Users/MR%20LAPTOP/pipwin/GDAL-3.4.3-cp310-cp310-win_amd64.whl#sha256=9893da5532a6c5dadc1fc3bc402ae29d161b7cf190d778f129da84d0e50baa29
idna==3.4
Js2Py==0.74
kombu==5.3.4
lxml==4.9.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pipwin==0.5.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
South==1.0.2
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unicodecsv==0.14.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
